0.5.1

+++ .vitepress/check.sh
@@ -1,13 +1,14 @@
 #!/bin/bash
 
 flag_sync=$1
+flag_post=$2
 
 echo checking package.json tsconfig.base.json
 FNS="package.json tsconfig.base.json"
 target=..
 [ -d $target ] || ( echo target not found; exit )
 for FN in $FNS; do
-	if [ -n "$flag_sync" ]; then
+	if [ -n "$flag_sync" -a -f "$FN" ]; then
 		cp -v $FN $target/$FN
 	else
 		echo not copying $FN
@@ -31,4 +32,11 @@ for FN in `find src -type f`; do
 		fi
 	fi
 done
+
+if [ -n "$flag_post" ]; then
+	sed -ibak 's/_assets/assets/' ../dist/client/app/utils.js
+	rm -f ../dist/client/app/utils.jsbak
+	sed -ibak 's/_assets/assets/' ../dist/node/build/render.js
+	rm -f ../dist/node/build/render.jsbak
+fi
